[FIX] project_task_materials_stock: Error when create a project from template. [FIX] project_task_materials_stock: Duplicate entries if user configure more than one stage to consume material

+++ project_task_materials_stock/models/project.py
@@ -50,24 +50,29 @@ class Task(models.Model):
 
     @api.multi
     def unlink_stock_move(self):
+        res = False
         moves = self.mapped('stock_move_ids')
-        moves.filtered(lambda r: r.state == 'assigned').do_unreserve()
-        moves.filtered(
-            lambda r: r.state in ['waiting', 'confirmed', 'assigned']
-        ).write({'state': 'draft'})
-        moves.unlink()
+        moves_done = moves.filtered(lambda r: r.state == 'done')
+        if not moves_done:
+            moves.filtered(lambda r: r.state == 'assigned').do_unreserve()
+            moves.filtered(
+                lambda r: r.state in ['waiting', 'confirmed', 'assigned']
+            ).write({'state': 'draft'})
+            res = moves.unlink()
+        return res
 
     @api.multi
     def write(self, vals):
+        res = super(Task, self).write(vals)
         for task in self:
-            res = super(Task, self).write(vals)
             if 'stage_id' in vals:
                 if task.stage_id.consume_material:
-                    task.material_ids.create_stock_move()
-                    task.material_ids.create_analytic_line()
+                    if not task.stock_move_ids:
+                        task.material_ids.create_stock_move()
+                        task.material_ids.create_analytic_line()
                 else:
-                    task.unlink_stock_move()
-                    task.analytic_line_ids.unlink()
+                    if task.unlink_stock_move():
+                        task.analytic_line_ids.unlink()
         return res
 
     @api.multi
